requirements.txt
transformers
sentence-transformers
faiss-cpu
torch
tqdm
scikit-learn
pandas
numpy
matplotlib
streamlit
gradio
pdfplumber
ipykernel
python-docx
nltk
